+++ app/src/main/res/layout/activity_view_meetups.xml
@@ -9,19 +9,14 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="uk.ac.tees.donut.squad.activities.ViewMeetups">
+    tools:context="uk.ac.tees.donut.squad.activities.ViewMeetups"
+    android:background="@color/cardview_dark_background">
 
     <android.support.v7.widget.RecyclerView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:id="@+id/viewMeetups_recyclerView"
         android:layout_alignParentStart="true"
-        android:layout_below="@+id/spinner" />
-
-    <Spinner
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentStart="true"
-        android:id="@+id/spinner" />
+        android:background="@color/colorPrimary"
+        android:padding="5dp"/>
 </RelativeLayout>
